refactor: extract MediaImageViewer from MediaModal for better separation of concerns
• Create new MediaImageViewer component handling all image display and navigation
• Move complete touch gesture system and visual feedback to MediaImageViewer
• Extract swipe animations, progress indicators, and navigation zones
• Simplify MediaModal to focus on modal structure and description editing
• Maintain 100% functionality while improving code organization
• Enable reusability of image viewer component across the application

Benefits:
- MediaModal reduced from ~500 to ~300 lines
- Clear separation between modal and image viewer responsibilities
- Easier testing and maintenance of focused components
- MediaImageViewer can be reused in other contexts
